Split vanilla and sod common
