+++ sim/common/includes.go
@@ -2,6 +2,9 @@ package common
 
 // Just import other directories, so importing common from elsewhere is enough.
 import (
+	_ "github.com/wowsims/sod/sim/common/vanilla"
+	_ "github.com/wowsims/sod/sim/common/vanilla/items_sets"
+
 	_ "github.com/wowsims/sod/sim/common/sod"
 	_ "github.com/wowsims/sod/sim/common/sod/crafted"
 	_ "github.com/wowsims/sod/sim/common/sod/item_effects"
