+++ sim/common/vanilla/caster_items.go
