+++ sim/common/vanilla/melee_trinkets.go
